Implements gh-3960, change size and alignment to be size_t
Size and alignment struct members must have same types as
return types of sizeof(T) and of alignof(T) respectively,
which are both `size_t`.

+++ c/parallel/src/merge_sort.cu
@@ -208,7 +208,7 @@ struct dynamic_merge_sort_policy_t
     return op.template Invoke<merge_sort_runtime_tuning_policy>(GetPolicy(device_ptx_version, key_size));
   }
 
-  int key_size;
+  uint64_t key_size;
 };
 
 struct merge_sort_kernel_source
